nibiru-temp/requirements.txt
fastapi
uvicorn[standard]
watchgod
redis[async]
pillow
